tests product

+++ src/main/java/ar/com/flexibility/examen/domain/service/impl/ProductServiceImpl.java
@@ -1,8 +1,10 @@
 package ar.com.flexibility.examen.domain.service.impl;
 
+import ar.com.flexibility.examen.domain.exception.GenericProductException;
 import ar.com.flexibility.examen.domain.model.Product;
 import ar.com.flexibility.examen.domain.repository.ProductRepository;
 import ar.com.flexibility.examen.domain.service.ProductService;
+import javassist.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,29 +18,49 @@ public class ProductServiceImpl implements ProductService {
 
 
     @Override
-    public List<Product> findAll() {
-        return productRepository.findAll();
-    }
-
-    @Override
-    public Product findOne(Long id){
-        return productRepository.findOne(id);
-    }
-
-    @Override
     public Boolean exists(Long id){
         return productRepository.exists(id);
     }
 
     @Override
-    public Product add(Product product) {
-        return productRepository.saveAndFlush(product);
+    public void deleteAll(){
+        productRepository.deleteAll();
     }
 
     @Override
-    public Product update(Product product) {
+    public List<Product> findAll() {
+        return productRepository.findAll();
+    }
 
-        Product productToPersist = productRepository.getOne(product.getId());
+    @Override
+    public Product findOne(Long id) throws NotFoundException {
+        Product product = productRepository.findOne(id);
+        if (product == null) {
+            throw new NotFoundException(String.format(PRODUCT_ID_NOT_EXIST, id));
+        }
+        return product;
+    }
+
+    @Override
+    public Product add(Product product) throws GenericProductException {
+        Product productAdded = productRepository.saveAndFlush(product);
+        if (productAdded == null) {
+            throw new GenericProductException(PRODUCT_ADDED_FAILED);
+        }
+        return productAdded;
+    }
+
+    @Override
+    public Product update(Product product) throws NotFoundException, GenericProductException {
+
+        Product productToPersist = productRepository.findOne(product.getId());
+
+        if (productToPersist == null) {
+            throw new NotFoundException(String.format(PRODUCT_ID_NOT_EXIST, product.getId()));
+
+        } else if (productToPersist.equals(product)){
+            throw new GenericProductException(PRODUCT_TO_UPDATE_WITHOUT_CHANGES);
+        }
 
         productToPersist.setDescription(product.getDescription());
         productToPersist.setPrice(product.getPrice());
@@ -47,9 +69,18 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Boolean delete(Long id) {
+    public void delete(Long id) throws NotFoundException, GenericProductException {
+
+        if (!exists(id)) {
+            throw new NotFoundException(String.format(PRODUCT_ID_NOT_EXIST, id));
+        }
+
         productRepository.delete(id);
-        return !exists(id);
+
+        if (exists(id)){
+            throw new GenericProductException(PRODUCT_DELETE_FAILED);
+        }
+
     }
 
 }
